requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
pywin32>=307
pandas>=1.3.5
numpy>=1.26.0
matplotlib>=3.5.1
plotly>=5.6.0
openpyxl>=3.0.9
dash>=2.3.0
dash-bootstrap-components>=1.0.2
python-dotenv>=0.20.0
